Conflicts:
	price.py

+++ price.py
@@ -16,8 +16,8 @@
 # along with Marketeer.  If not, see <http://www.gnu.org/licenses/>.
 
 from time import time
-from decimal import *
 import sqlite3
+from decimal import *
 
 class Price():
     '''
@@ -102,8 +102,8 @@ class Store():
                 security TEXT NOT NULL,
                 currency TEXT NOT NULL,
                 timestamp INTEGER NOT NULL,
-                bid FLOAT NOT NULL,
-                offer FLOAT NOT NULL,
+                bid DECIMAL(18, 6) NOT NULL,
+                offer DECIMAL(18, 6) NOT NULL,
                 PRIMARY KEY (exchange, security, currency, timestamp))""")
 
         c.execute("""CREATE TABLE IF NOT EXISTS price_data (
